Rewrite

+++ index.php
@@ -7,33 +7,47 @@ $result = $conn->query("SELECT * FROM users");
 
 <head>
     <title>CRUD PHP MySQL</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" type="text/css" href="bootstrap533/css/bootstrap.css">
+    <link rel="stylesheet" type="text/css" href="css/styleku.css">
+    <script src="bootstrap533/jquery/3.3.1/jquery-3.3.1.js"></script>
+    <script src="bootstrap533/js/bootstrap.js"></script>
 </head>
 
 <body>
-    <h2>Daftar Pengguna</h2>
-    <a href="create.php">Tambah Pengguna</a>
-    <table border="1">
-        <tr>
-            <th>ID</th>
-            <th>Nama</th>
-            <th>Email</th>
-            <th>Password</th>
-            <th>Aksi</th>
-        </tr>
-        <?php while ($row = $result->fetch_assoc()): ?>
-            <tr>
-                <td><?= $row['id'] ?></td>
-                <td><?= $row['name'] ?></td>
-                <td><?= $row['email'] ?></td>
-                <td><?= $row['passw'] ?></td>
-                <td>
-                    <a href="edit.php?id=<?= $row['id'] ?>">Edit</a> |
-                    <a href="delete.php?id=<?= $row['id'] ?>" onclick="return
-confirm('Hapus data?')">Hapus</a>
-                </td>
-            </tr>
-        <?php endwhile; ?>
-    </table>
+    <div class="utama">
+        <h2 class="text-center">Daftar Pengguna</h2>
+        <span class="float-left">
+            <a class="btn btn-success" href="create.php">Tambah Pengguna</a>
+        </span>
+        <div class="container">
+            <table class="table table-hover">
+                <thead class="thead-light">
+                    <tr>
+                        <th>ID</th>
+                        <th>Nama</th>
+                        <th>Email</th>
+                        <th>Password</th>
+                        <th>Aksi</th>
+                    </tr>
+
+                    <?php while ($row = $result->fetch_assoc()): ?>
+                        <tr>
+                            <td><?= $row['id'] ?></td>
+                            <td><?= $row['name'] ?></td>
+                            <td><?= $row['email'] ?></td>
+                            <td><?= $row['passw'] ?></td>
+                            <td>
+                                <a href="edit.php?id=<?= $row['id'] ?>">Edit</a> |
+                                <a href="delete.php?id=<?= $row['id'] ?>" onclick="return confirm('Hapus data?')">Hapus</a>
+                            </td>
+                        </tr>
+                    <?php endwhile; ?>
+                </thead>
+            </table>
+        </div>
+    </div>
 </body>
 
 </html>
